Sacred stat generation changes
The algorithm changed considerably so amounts may be a bit strange.
Changing stats too far away from race/pose basics should now be way more
rare and way more expensive for "stat budget".

This closes #454

+++ src/nationGen/magic/MageGenerator.java
@@ -1509,7 +1509,7 @@ public class MageGenerator extends TroopGenerator {
 			// If all mages of tier have distinguishing paths and either luck or no common paths, separate filters are rolled. 
 			if(different)
 			{
-				int maxpower = 0;
+				double maxpower = 0;
 				for(Unit u : mages)
 				{
 					List<Filter> givenFilters = this.getPossibleFiltersByPaths(actualFilters, MageGenerator.findDistinguishingPaths(u, mages), strict);
